% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_vikor.R, R/probne.R
\name{.get_final_weights}
\alias{.get_final_weights}
\title{Process Weights for Fuzzy MCDM}
\usage{
.get_final_weights(decision_mat, weights, bwm_criteria, bwm_best, bwm_worst)

.get_final_weights(decision_mat, weights, bwm_criteria, bwm_best, bwm_worst)
}
\description{
Internal helper to handle either explicit fuzzy weights or BWM calculation.

Internal helper to handle either explicit fuzzy weights or BWM calculation.
}
\keyword{internal}
