% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_ranking.R
\name{.oblicz_ranking_dominacji}
\alias{.oblicz_ranking_dominacji}
\title{Teoria dominacji dla rankingu}
\usage{
.oblicz_ranking_dominacji(r1, r2, r3)
}
\arguments{
\item{r1}{Wektor numeryczny rang metody 1.}

\item{r2}{Wektor numeryczny rang metody 2.}

\item{r3}{Wektor numeryczny rang metody 3.}
}
\value{
Wektor numeryczny zawierający końcowy ranking alternatyw.
}
\description{
Funkcja pomocnicza wyznaczająca ranking konsensusu na podstawie reguły
większości. Iteracyjnie sprawdza, która alternatywa najczęściej wygrywa
na danej pozycji w rankingach cząstkowych.
}
\keyword{internal}
