% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.parsuj_skladnie_mcda}
\alias{.parsuj_skladnie_mcda}
\title{Parsowanie składni MCDA}
\usage{
.parsuj_skladnie_mcda(skladnia)
}
\arguments{
\item{skladnia}{Ciąg znaków opisujący składnię modelu MCDA.}
}
\value{
Lista, w której nazwy elementów odpowiadają kryteriom,
a wartości są wektorami nazw zmiennych przypisanych do danego kryterium.
}
\description{
Funkcja pomocnicza interpretująca model zapisany jako ciąg znaków
w postaci np. `"Kryterium =~ zmienna1 + zmienna2"`.
Zwraca listę mapującą nazwy kryteriów na odpowiadające im zmienne.
}
\keyword{internal}
