% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wagi_pomocnicze.R
\name{.pobierz_finalne_wagi}
\alias{.pobierz_finalne_wagi}
\title{Wewnętrzny procesor wag}
\usage{
.pobierz_finalne_wagi(
  macierz,
  wagi = NULL,
  bwm_kryteria = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL
)
}
\description{
Decyduje, skąd pobrać wagi kryteriów: z wektora podanego ręcznie
albo z metody BWM.
}
\keyword{internal}
