% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.rozmyj_wektor}
\alias{.rozmyj_wektor}
\title{Wewnętrzna funkcja rozmywiająca}
\usage{
.rozmyj_wektor(wektor)
}
\arguments{
\item{wektor}{Wektor numeryczny.}
}
\value{
Macierz trójkolumnowa zawierająca wartości `l`, `m`, `u`.
}
\description{
Zamienia liczbę rzeczywistą na trójkątną liczbę rozmytą (TFN).
Dla wartości `x` tworzona jest trójka `(l, m, u)`, gdzie:
`l = x - 1`, `m = x`, `u = x + 1`,
z ograniczeniem do przedziału 1-9.
}
\keyword{internal}
