% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.skaluj_do_saaty}
\alias{.skaluj_do_saaty}
\title{Wewnętrzny skaler Saaty'ego}
\usage{
.skaluj_do_saaty(wektor)
}
\arguments{
\item{wektor}{Wektor numeryczny.}
}
\value{
Wektor numeryczny przeskalowany do przedziału 1-9.
}
\description{
Przekształca dowolną skalę, np. Likert 1-5 lub wartości ciągłe,
do skali Saaty'ego 1-9.
}
\keyword{internal}
