% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_multimoora_wynik}
\alias{plot.rozmyty_multimoora_wynik}
\title{Mapa porownania skladnikow MULTIMOORA}
\usage{
\method{plot}{rozmyty_multimoora_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_multimoora_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Wizualizacja porownuje dwa glowne skladniki wyniku MULTIMOORA:
Ratio System oraz Full Multiplicative Form.
Wielkosc punktu zalezy od koncowej sumy rang,
a kolor od pozycji w rankingu koncowym.
}
