% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja.R
\name{plot.rozmyty_topsis_wynik}
\alias{plot.rozmyty_topsis_wynik}
\title{Mapa efektywnosci TOPSIS}
\usage{
\method{plot}{rozmyty_topsis_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_topsis_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Pokazuje odleglosc od rozwiazania idealnego.
Os X: dystans od antywzorca (D-).
Os Y: dystans do wzorca (D+).
}
