% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykresy.R
\name{plot.rozmyty_vikor_wynik}
\alias{plot.rozmyty_vikor_wynik}
\title{Mapa strategiczna VIKOR}
\usage{
\method{plot}{rozmyty_vikor_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy `rozmyty_vikor_wynik`.}

\item{...}{Dodatkowe argumenty.}
}
\description{
Wizualizacja typu cIPMA.
Os X: efektywnosc grupowa (odwrocone S).
Os Y: ryzyko / zal (R).
Wielkosc bąbla: sila kompromisu zależna od Q.
}
