% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyty_multimoora.R
\name{rozmyty_multimoora}
\alias{rozmyty_multimoora}
\title{Rozmyty MULTIMOORA}
\usage{
rozmyty_multimoora(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL,
  bwm_kryteria = NULL
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz decyzyjna, gdzie wiersze oznaczają
alternatywy, a kolumny kryteria.}

\item{typy_kryteriow}{Wektor typów kryteriów: \code{"max"} albo \code{"min"}.}

\item{wagi}{Opcjonalny wektor wag kryteriów.}

\item{bwm_najlepsze}{Opcjonalny wektor Best-to-Others dla BWM.}

\item{bwm_najgorsze}{Opcjonalny wektor Others-to-Worst dla BWM.}

\item{bwm_kryteria}{Opcjonalny wektor nazw kryteriów dla BWM.}
}
\value{
Obiekt klasy \code{"rozmyty_multimoora_wynik"}.
}
\description{
Wyznacza ranking alternatyw metodą MULTIMOORA dla macierzy decyzyjnej.
}
