% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyty_topsis.R
\name{rozmyty_topsis}
\alias{rozmyty_topsis}
\title{Rozmyta metoda TOPSIS}
\usage{
rozmyty_topsis(
  macierz_decyzyjna,
  typy_kryteriow,
  wagi = NULL,
  bwm_kryteria = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz o wymiarach \eqn{m x 3n}.}

\item{typy_kryteriow}{Wektor znakowy: \code{"max"} dla kryteriów zyskowych
oraz \code{"min"} dla kryteriów kosztowych.}

\item{wagi}{Opcjonalny wektor wag.}

\item{bwm_kryteria}{Opcjonalne nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{Opcjonalny wektor Best-to-Others.}

\item{bwm_najgorsze}{Opcjonalny wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_topsis_wynik` z rankingiem.
}
\description{
Implementacja Fuzzy TOPSIS. Oblicza odległość od rozwiązania idealnego
i antyidealnego.
}
