% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmyty_vikor.R
\name{rozmyty_vikor}
\alias{rozmyty_vikor}
\title{Rozmyta metoda VIKOR}
\usage{
rozmyty_vikor(
  macierz_decyzyjna,
  typy_kryteriow,
  v = 0.5,
  wagi = NULL,
  bwm_kryteria = NULL,
  bwm_najlepsze = NULL,
  bwm_najgorsze = NULL
)
}
\arguments{
\item{macierz_decyzyjna}{Macierz o wymiarach \eqn{m x 3n}.}

\item{typy_kryteriow}{Wektor znakowy: \code{"max"} dla kryteriów zyskowych
oraz \code{"min"} dla kryteriów kosztowych.}

\item{v}{Waga strategii większości kryteriów. Domyślnie `0.5`.}

\item{wagi}{Opcjonalny wektor wag.}

\item{bwm_kryteria}{Opcjonalne nazwy kryteriów dla BWM.}

\item{bwm_najlepsze}{Opcjonalny wektor Best-to-Others.}

\item{bwm_najgorsze}{Opcjonalny wektor Others-to-Worst.}
}
\value{
Obiekt klasy `rozmyty_vikor_wynik`.
}
\description{
Metoda kompromisowa VIKOR. Oblicza wskaźniki S, R oraz Q.
}
