% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{tabela_apa}
\alias{tabela_apa}
\title{Generowanie tabeli APA}
\usage{
tabela_apa(x, tytul = NULL)
}
\arguments{
\item{x}{Obiekt wynikowy z funkcji pakietu.}

\item{tytul}{Opcjonalny tytul tabeli.}
}
\value{
Obiekt klasy `flextable`.
}
\description{
Funkcja przeksztalca wyniki analizy MCDA
(TOPSIS, VIKOR, MULTIMOORA, meta-ranking)
w sformatowana tabele zgodna ze standardem APA.
}
