% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zastosuj_profil_uzytkownika.R
\name{zastosuj_profil_uzytkownika}
\alias{zastosuj_profil_uzytkownika}
\title{Personalizacja wag kryteriów według profilu użytkownika}
\usage{
zastosuj_profil_uzytkownika(wagi, profil)
}
\arguments{
\item{wagi}{Numeryczny wektor wag kryteriów. Wektor musi posiadać nazwy
odpowiadające kryteriom decyzyjnym.}

\item{profil}{Nazwa profilu użytkownika (np. "zestresowany_student",
"student_analityczny", "student_oszczedny").}
}
\value{
Numeryczny wektor wag po modyfikacji i normalizacji.
}
\description{
Funkcja modyfikuje bazowy wektor wag kryteriów na podstawie
wybranego profilu użytkownika, a następnie dokonuje ich normalizacji.
Pozwala to uwzględnić preferencje różnych typów użytkowników
w procesie wspomagania decyzji.
}
\examples{
wagi <- c(
  dokladnosc_faz_snu = 0.2,
  analiza_statystyk = 0.15,
  ux_latwosc = 0.1,
  dane_fizjologiczne = 0.1,
  integracja_urzadzenia = 0.1,
  funkcje_dodatkowe = 0.1,
  zuzycie_baterii = 0.15,
  cena_subskrypcja = 0.1
)

zastosuj_profil_uzytkownika(wagi, "zestresowany_student")

}
